Game starts and is playable
